refactor(infoBox): info 컴포넌트 분리
