+++ src/components/maps/index.ts
@@ -1,4 +1,4 @@
 export { default as MapContainer } from './MapContainer';
 export { default as Map } from './Map';
 export { default as Marker } from './Marker';
-export { default as InfoBox } from './InfoBox';
+export { default as InfoContainer } from './infos/InfoContainer';
